requirements.txt
astroid==2.3.2
Click==7.0
colorama==0.4.1
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
pylint==2.4.3
six==1.12.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.1
